requirements.txt
django
Pillow
python-Levenshtein
requests